+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 ## 8.13.0-SNAPSHOT
 - Updated Chinese (Traditional) translation
+- Added option to log ChestSort actions in a log file (default: false)
 
 ## 8.12.2
 - Fixes NullPointerException when using EssentialsX' /ec command in Minecraft 1.12